requirements.txt
boto3>=1.9,<2
apache-airflow
psycopg2-binary
celery
python-dotenv
